clk: qcom: gcc-x1e80100: Do not register gpu_cfg_ahb always-on clock
On probe, we enable the gpu_cfg_ahb clock expecting it to remain always
on beyond that point, but when clk_disable_unused late_initcall is
called, since the clock is not attached to any consumer, the clock gets
disabled. Avoid that by not registering it with the clock framework.

Signed-off-by: Abel Vesa <abel.vesa@linaro.org>