UBUNTU: [Packaging] linux-bpf-dev: Restructure packaging
Building the linux-bpf-dev package requires bpftool and vmlinux which are
produced in different build steps. The current build/install order looks
like this:

1. build-flavor
   - build vmlinux

2. install-flavor
   - copy vmlinux to tmp/vmlinux
   - remove the flavor build directory

3. build-tools
   - build bpftool
   - bpftool tmp/vmlinux > vmlinux.h
   - remove tmp/vmlinux

4. install-tools
   - copy vmlinux.h to package directory
...

To fix out-of-space issues in LP, we have to remove the flavor build
directory at the end of the install-flavor step which means vmlinux
is no longer around afterwards, hence the ugly tmp/vmlinux dance.
Also, vmlinux.h is really a flavor build artifact so should be generated
in the build-flavor rather than the build-tools step.

Clean this up some by reording. Build the tools first and then produce
per-flavor linux.h and install it in the following step.

1. build-tools
   - build bpftool

2. install-tools

3. build-flavor
   - build vmlinux
   - bpftool vmlinux > vmlinux.h

4. install-flavor
   - copy vmlinux.h to package directory
   - remove the flavor build directory
...

Also drop the build-arch dependency from the stamp-build-perarch target
to prevent a circular dependency.

Note that this reordering is also necessary for an upcoming commit where
we want to install the tools into flavored packages.

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>