platform: arm64: Add driver for EC found in most X1E laptops
Currently it features only reporting that the AP is going to suspend,
which results in keyboard backlight turning off and the power LED
slowly blinking on the Lenovo Yoga Slim 7x.

Honor Magicbook Art 14 and Lenovo Yoga Slim 7x are known to have
firmware with extensions which would need appropriate handling.
For reverse engineering the firmware on them I have written a Rust
utility:

https://github.com/Maccraft123/it8987-qcom-tool.git

Signed-off-by: Maya Matuszczyk <maccraft123mc@gmail.com>