UBUNTU: SAUCE: apparmor4.0.0 [41/53]: apparmor: transition from a list of rules to a vector of rules
BugLink: http://bugs.launchpad.net/bugs/2028253

The set of rules on a profile is not dynamically extended, instead
if a new ruleset is needed a new version of the profile is created.
This allows us to use a vector of rules instead of a list, slightly
reducing memory usage and simplifying the code some.

Signed-off-by: John Johansen <john.johansen@canonical.com>