arm64: dts: qcom: x1-crd: mark l12b and l15b always-on
The l12b and l15b supplies are used by components that are not (fully)
described (and some never will be) and must never be disabled.

Mark the regulators as always-on to prevent them from being disabled,
for example, when consumers probe defer or suspend.

Fixes: bd50b1f5b6f3 ("arm64: dts: qcom: x1e80100: Add Compute Reference Device")
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>

amended for commonized variant of X1-CRD with x1e80100 and x1p42100 variants