arm64: dts: qcom: x1e80100-dell-xps13-9345: mark l12b and l15b always-on
The l12b and l15b supplies are used by components that are not (fully)
described (and some never will be) and must never be disabled.

Mark the regulators as always-on to prevent them from being disabled,
for example, when consumers probe defer or suspend.

Note that these supplies currently have no consumers described in
mainline.

Fixes: f5b788d0e8cd ("arm64: dts: qcom: Add support for X1-based Dell XPS 13 9345")
Cc: stable@vger.kernel.org	# 6.13
Cc: Aleksandrs Vinarskis <alex.vinarskis@gmail.com>
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>