arm64: dts: qcom: x1p42100-lenovo-thinkbook-16: enable pwm backlight
after some more analysis, it should be the normal PWM backlight, with a PWM
frequency of 2kHz. The PWM is in the pm8550, also the controlling GPIOs

Derived with Aleksandrs Vinarskis <alex.vinarskis@gmail.com> from the dsdt,
compared against HP Omnibook dsdt.

The definition is in the PCG2 rom table for the CMN162B display.

</Group>
<Group id='Backlight Configuration'>
<BacklightType>1</BacklightType>
<BacklightPmicControlType>3</BacklightPmicControlType>
<BacklightPmicId>1</BacklightPmicId>
<BacklightPmicBankSelect>3</BacklightPmicBankSelect>
<BacklightPmicPWMSizeinBits>9</BacklightPmicPWMSizeinBits>
<BacklightPMICPWMFrequency>2000</BacklightPMICPWMFrequency>
<BrightnessMaxLuminance>220000</BrightnessMaxLuminance>
<BrightnessMinLuminance>100000</BrightnessMinLuminance>
<BacklightSteps>100</BacklightSteps>
<BrightnessRangeLevel0>0 0 1000 2</BrightnessRangeLevel0>
<AdaptiveBrightnessFeature>1</AdaptiveBrightnessFeature>
<BacklightDefault>80</BacklightDefault>
<CABLEnable>True</CABLEnable>
<LTMEnable>1</LTMEnable>
</Group>

BacklightPmicID and BacklightPmicBankSelect appear to point to pm8550,
while on the Omnibook it is actually the pmk8550 (ID 0).

Signed-off-by: Jens Glathe <jens.glathe@oldschoolsolutions.biz>