arm64: dts: qcom: x1p42100-lenovo-thinkbook-16: enable camera on usb_2
Enabled and confirmed the camera, on the usb_2 controller

ID 5986:1198 Bison Electronics Inc. Integrated Camera

For it to be powered gpio44 needs to be unreserved and used for
vreg_cam_5p0, which of course is pure guesswork. But you can't
argue with succes.

Signed-off-by: Jens Glathe <jens.glathe@oldschoolsolutions.biz>
Co-developed by: Aleksandrs Vinarskis <alex.vinarskis@gmail.com>