arm64: dts: qcom: x1p42100-tb16: fix gpio-sbu-mux gpio assignments
Fix the asssignment of enable-gpios and select-gpios for usb-1-ss0-sbu-mux
and usb-1-ss1-sbu-mux. A look into the UCS0 structure in the dsdt table revealed different
assignment than on the HP Omnibook X14. The lines for /oe and /sel are interchanged.

With this change dp Altmode works on all orientations on the Thinkbook 16.

Signed-off-by: Jens Glathe <jens.glathe@oldschoolsolutions.biz>